refactor: remove find-by-id-and-company-id usecase and move it logic to delete-by-id usecase

+++ src/data/usecases/delete-court-by-id/delete-court-by-id.ts
@@ -1,11 +1,19 @@
-import { Court, DeleteCourtByIdRepository, IDeleteCourtById } from './delete-court-by-id.protocols'
+import { NotFoundError } from '@/domain/errors/not-found-error'
+import { Court, DeleteCourtByIdRepository, FindCourtByIdAndCompanyIdRepository, IDeleteCourtById } from './delete-court-by-id.protocols'
 
 export class DeleteCourtById implements IDeleteCourtById {
   constructor (
-    private readonly deleteCourt: DeleteCourtByIdRepository
+    private readonly deleteCourt: DeleteCourtByIdRepository,
+    private readonly findCourt: FindCourtByIdAndCompanyIdRepository
   ) {}
 
-  async deleteById (courtId: string): Promise<Court> {
+  async deleteById (courtId: string, companyId: string): Promise<Court | Error> {
+    const court = await this.findCourt.findByIdAndCompanyId(courtId, companyId)
+
+    if(!court) {
+      return new NotFoundError('court')
+    }
+
     return await this.deleteCourt.deleteById(courtId)
   }
 }
